cambiar codigo a long

+++ SmartMarkt/SmartMarkt/Product.cs
@@ -11,6 +11,6 @@ namespace SmartMarkt
         public int Id { get; set; }
         public string Name { get; set; }
         public string Address { get; set; }
-        public int BarCode { get; set; }
+        public long BarCode { get; set; }
     }
 }
